% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.R
\name{addThings}
\alias{addThings}
\title{Example function.}
\usage{
addThings(x, y)
}
\arguments{
\item{x}{First input}

\item{y}{Second input}
}
\value{
Solver output
}
\description{
This function will not work but illustrates the use of gurobi and rglpk
}
